feat: move buildDeltaMapping to somewhere it can be imported
